Do Minh Hieu

+++ inversion_algorithm.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 
-int p, a, u, v, x1, x2, r, q, x;
+unsigned long long p, a, u, v, x1, x2, r, q, x;
 
 int inverse(){
 	u = a % p;
@@ -19,14 +19,14 @@ int inverse(){
 	return x1;
 }
 void display(){
-	printf("\nd = %d", x1);
-	printf("\n%d^-1 = %d (mod %d)", a, x1, p);
-	printf("\n%d.%d = 1 (mod %d)", a, x1, p);
+	printf("\nd = %lld", x1);
+	printf("\n%lld^-1 = %lld (mod %lld)", a, x1, p);
+	printf("\n%lld.%lld = 1 (mod %lld)", a, x1, p);
 }
 int main(){
 	printf("Phep tinh nghich dao a^-1 = d (mod p) hay a.d = 1 (mod p)\n");
 	printf("\nEnter p, a: ");
-	scanf("%d %d", &p, &a);
+	scanf("%lld %lld", &p, &a);
 	inverse();
 	display();
 	return 0;
